World Generator 3.2
Finish the world generator for world boundary. The strategy is that the
world is composed of cubes. Each cubes is distinguished represented by
its center position, and these cubes' information are stored in a
ArrayList. Whenever a player is in one cube, such function will be
triggered to check whether this cube is surrounded entirely or not. If
not, new cube bound will be added to the ArrayList.